backend/requirements.txt
fastapi
uvicorn
google-genai
weasyprint
pydantic
jinja2
python-multipart
python-dotenv
requests
httpx
